interface/cli/starport: made root cmd newable
in order to:
* reduce package level globals,
* increase configurability of the root command by outside caller where such options can be passed as an argument to `starportcmd.New()`.
