+++ interface/cli/starport/cmd/typed.go
@@ -1,4 +1,4 @@
-package cmd
+package starportcmd
 
 import (
 	"context"
