+++ interface/cli/starport/main.go
@@ -1,7 +1,44 @@
 package main
 
-import "github.com/tendermint/starport/interface/cli/starport/cmd"
+import (
+	"fmt"
+	"os"
+
+	starportcmd "github.com/tendermint/starport/interface/cli/starport/cmd"
+	"github.com/tendermint/starport/pkg/analyticsutil"
+)
 
 func main() {
-	cmd.Execute()
+	defer func() {
+		if r := recover(); r != nil {
+			addMetric(Metric{
+				Err: fmt.Errorf("%s", r),
+			})
+			analyticsc.Close()
+			fmt.Println(r)
+			os.Exit(1)
+		}
+	}()
+	analyticsc = analyticsutil.New(analyticsEndpoint, analyticsKey)
+	// TODO add version of new installation.
+	name, hadLogin := prepLoginName()
+	analyticsc.Login(name, "todo-version")
+	if !hadLogin {
+		addMetric(Metric{
+			Login:          name,
+			IsInstallation: true,
+		})
+	}
+	if len(os.Args) > 1 && os.Args[1] == "serve" {
+		addMetric(Metric{})
+	}
+	err := starportcmd.New().Execute()
+	addMetric(Metric{
+		Err: err,
+	})
+	analyticsc.Close()
+	if err != nil {
+		fmt.Println(err)
+		os.Exit(1)
+	}
 }
